# api/docs/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  description: "API for managing sandboxes, organizations, billing, and hosts"
  title: Fluid API
  version: "1.0"
servers:
- url: //localhost:8081/v1
paths:
...
  /orgs/{slug}/sandboxes:
    get:
      description: List all sandboxes in the organization
      parameters:
      - description: Organization slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Internal Server Error
      security:
      - CookieAuth: []
      summary: List sandboxes
      tags:
      - Sandboxes
    post:
      description: Create a new sandbox in the organization from a source VM or base
        image
      parameters:
      - description: Organization slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/orchestrator.CreateSandboxRequest"
        description: Sandbox configuration
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/store.Sandbox"
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Internal Server Error
      security:
      - CookieAuth: []
      summary: Create sandbox
      tags:
      - Sandboxes
      x-codegen-request-body-name: request
  /orgs/{slug}/sandboxes/{sandboxID}:
    delete:
      description: Destroy a sandbox and release its resources
      parameters:
      - description: Organization slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      - description: Sandbox ID
        in: path
        name: sandboxID
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Internal Server Error
      security:
      - CookieAuth: []
      summary: Destroy sandbox
      tags:
      - Sandboxes
    get:
      description: Get sandbox details by ID
      parameters:
      - description: Organization slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      - description: Sandbox ID
        in: path
        name: sandboxID
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/store.Sandbox"
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Not Found
      security:
      - CookieAuth: []
      summary: Get sandbox
      tags:
      - Sandboxes
  /orgs/{slug}/sandboxes/{sandboxID}/commands:
    get:
      description: List all commands executed in a sandbox
      parameters:
      - description: Organization slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      - description: Sandbox ID
        in: path
        name: sandboxID
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Internal Server Error
      security:
      - CookieAuth: []
      summary: List commands
      tags:
      - Sandboxes
  /orgs/{slug}/sandboxes/{sandboxID}/ip:
    get:
      description: Get the IP address of a sandbox
      parameters:
      - description: Organization slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      - description: Sandbox ID
        in: path
        name: sandboxID
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Not Found
      security:
      - CookieAuth: []
      summary: Get sandbox IP
      tags:
      - Sandboxes
  /orgs/{slug}/sandboxes/{sandboxID}/run:
    post:
      description: Execute a command in a sandbox
      parameters:
      - description: Organization slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      - description: Sandbox ID
        in: path
        name: sandboxID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/orchestrator.RunCommandRequest"
        description: Command to run
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/store.Command"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Internal Server Error
      security:
      - CookieAuth: []
      summary: Run command
      tags:
      - Sandboxes
      x-codegen-request-body-name: request
  /orgs/{slug}/sandboxes/{sandboxID}/snapshot:
    post:
      description: Create a snapshot of a sandbox
      parameters:
      - description: Organization slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      - description: Sandbox ID
        in: path
        name: sandboxID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/orchestrator.SnapshotRequest"
        description: Snapshot details
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/orchestrator.SnapshotResponse"
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Internal Server Error
      security:
      - CookieAuth: []
      summary: Create snapshot
      tags:
      - Sandboxes
      x-codegen-request-body-name: request
  /orgs/{slug}/sandboxes/{sandboxID}/start:
    post:
      description: Start a stopped sandbox
      parameters:
      - description: Organization slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      - description: Sandbox ID
        in: path
        name: sandboxID
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Internal Server Error
      security:
      - CookieAuth: []
      summary: Start sandbox
      tags:
      - Sandboxes
  /orgs/{slug}/sandboxes/{sandboxID}/stop:
    post:
      description: Stop a running sandbox
      parameters:
      - description: Organization slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      - description: Sandbox ID
        in: path
        name: sandboxID
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Internal Server Error
      security:
      - CookieAuth: []
      summary: Stop sandbox
      tags:
      - Sandboxes
...
components:
  schemas:
    orchestrator.CreateSandboxRequest:
      properties:
        agent_id:
          type: string
        base_image:
          type: string
        memory_mb:
          type: integer
        name:
          type: string
        network:
          type: string
        org_id:
          type: string
        source_vm:
          type: string
        ttl_seconds:
          type: integer
        vcpus:
          type: integer
      type: object
...
    orchestrator.RunCommandRequest:
      properties:
        command:
          type: string
        env:
          additionalProperties:
            type: string
          type: object
        timeout_seconds:
          type: integer
      type: object
...
    orchestrator.SnapshotRequest:
      properties:
        name:
          type: string
      type: object
    orchestrator.SnapshotResponse:
      example:
        snapshot_name: snapshot_name
        snapshot_id: snapshot_id
        sandbox_id: sandbox_id
        created_at: created_at
      properties:
        created_at:
          type: string
        sandbox_id:
          type: string
        snapshot_id:
          type: string
        snapshot_name:
          type: string
      type: object
...
    rest.swaggerError:
      example:
        code: 0
        details: details
        error: error
      properties:
        code:
          type: integer
        details:
          type: string
        error:
          type: string
      type: object
...
    store.Command:
      example:
        duration_ms: 0
        stdout: stdout
        exit_code: 6
        sandbox_id: sandbox_id
        started_at: started_at
        id: id
        stderr: stderr
        command: command
        ended_at: ended_at
      properties:
        command:
          type: string
        duration_ms:
          type: integer
        ended_at:
          type: string
        exit_code:
          type: integer
        id:
          type: string
        sandbox_id:
          type: string
        started_at:
          type: string
        stderr:
          type: string
        stdout:
          type: string
      type: object
    store.Sandbox:
      example:
        agent_id: agent_id
        tap_device: tap_device
        ttl_seconds: 6
        created_at: created_at
        ip_address: ip_address
        vcpus: 1
        deleted_at: deleted_at
        host_id: host_id
        base_image: base_image
        updated_at: updated_at
        mac_address: mac_address
        org_id: org_id
        name: name
        memory_mb: 0
        bridge: bridge
        id: id
        state: CREATING
        source_vm: source_vm
      properties:
        agent_id:
          type: string
        base_image:
          type: string
        bridge:
          type: string
        created_at:
          type: string
        deleted_at:
          type: string
        host_id:
          type: string
        id:
          type: string
        ip_address:
          type: string
        mac_address:
          type: string
        memory_mb:
          type: integer
        name:
          type: string
        org_id:
          type: string
        source_vm:
          type: string
        state:
          $ref: "#/components/schemas/store.SandboxState"
        tap_device:
          type: string
        ttl_seconds:
          type: integer
        updated_at:
          type: string
        vcpus:
          type: integer
      type: object
    store.SandboxState:
      enum:
      - CREATING
      - RUNNING
      - STOPPED
      - DESTROYED
      - ERROR
      type: string
      x-enum-varnames:
      - SandboxStateCreating
      - SandboxStateRunning
      - SandboxStateStopped
      - SandboxStateDestroyed
      - SandboxStateError
...